:bug: Consider partially migrated Bibles
Fixed a bug that prevented the merging of
markings, if one backup had already been
migrated to the Study Edition of the Revised
New World Translation, while the other one
still used the Standard Edition.
If JWLM detects this, it will also migrate
all the markings of the other backup to the
Study Edition.

Also see https://github.com/AndreasSko/go-jwlm/pull/46