Optimize detection for migration to nwtsty
In some situations it could happen that the merger
was not able to detect that one backup had already
been migrated to the Study Edition of the Revised
New World Translation, while the other one still was
using the Regular Edition. This could lead to errors
like:

> Could not create SQLite database for exporting: [...] UNIQUE constraint failed: UserMark.UserMarkGuid

This version improved the detection of such cases,
preventing those kinds of errors in the future.

More details: https://github.com/AndreasSko/go-jwlm/pull/108