Wait for download of backup file
Files stored on third-party file hosters (like OneDrive) are not
necessarily pre-download when accessed for the first time. This can
result in Not-Found errors.

This commit adjusts the `importBackup` logic to wait for up to 60
seconds for the file to be downloaded before continuing with the import.
This should fix common error reports about "no such file or directory".