Support downloading "non-ubiquitous" files(?)
Some users recently started to report
getting the following error message:

> The file couldn’t be opened because you don’t have permission to view it.

During debugging I noticed that
on iOS 18.4 files stored on OneDrive
don't seem to be considered
ubiquitous items, which triggers
that issue. I haven't seen the problem
with iOS 18.5, but just to be sure
we are now also supporting downloading
files that are not ubiquitous.

Also used the change to improve
capturing erros for Sentry.